refactor(routers): create routers without alamofire
